extract rpc-responses-common crate
