+++ crates/hash/src/lib.rs
@@ -8,11 +8,11 @@ use solana_program::hash::{
 };
 use solders_macros::{common_methods, pyhash, richcmp_full};
 
-use solders_traits_core::{
-    impl_display, pybytes_general_via_slice, CommonMethodsCore, PyFromBytesGeneral,
-    PyHash, RichcmpFull,
-};
 use solders_traits::handle_py_err;
+use solders_traits_core::{
+    impl_display, pybytes_general_via_slice, CommonMethodsCore, PyFromBytesGeneral, PyHash,
+    RichcmpFull,
+};
 
 #[pyclass(module = "solders.hash", subclass)]
 /// A SHA-256 hash, most commonly used for blockhashes.
