+++ crates/transaction/src/lib.rs
@@ -14,11 +14,11 @@ use solana_sdk::{
 };
 use solders_macros::{common_methods, richcmp_eq_only, EnumIntoPy};
 use solders_pubkey::{convert_optional_pubkey, Pubkey};
+use solders_traits::handle_py_err;
 use solders_traits_core::{
     impl_display, py_from_bytes_general_via_bincode, pybytes_general_via_bincode,
     CommonMethodsCore, RichcmpEqualityOnly,
 };
-use solders_traits::handle_py_err;
 
 use solders_hash::Hash as SolderHash;
 use solders_instruction::{convert_instructions, CompiledInstruction, Instruction};
